pyproject.toml
[tool.poetry]
name = "linc"
version = "0.1.0"
description = "Paper implementation of LINC"
authors = ["menamerai <rphananh@gmail.com>", "arnavkomaragiri <komaraac@mail.uc.edu>"]
readme = "README.md"
# package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
transformers = "^4.38.2"
datasets = "^2.18.0"
nltk = "^3.8.1"
matplotlib = "^3.8.3"
torch = "^2.1.0"
accelerate = "^0.28.0"
diskcache = "^5.6.3"
google-generativeai = "^0.4.1"
python-dotenv = "^1.0.1"
cohere = "^4.56"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.2"
black = "^24.2.0"
isort = "^5.13.2"

[tool.poetry.group.gpu-win.dependencies]
torch = { url = "https://download.pytorch.org/whl/cu121/torch-2.2.1%2Bcu121-cp310-cp310-win_amd64.whl" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
